fix: automatically remove define related configuration
