+++ packages/vitest/src/node/plugins/utils.ts
@@ -1,6 +1,6 @@
 import { builtinModules } from 'node:module'
 import { version as viteVersion } from 'vite'
-import type { DepOptimizationOptions } from 'vite'
+import type { DepOptimizationOptions, UserConfig as ViteConfig } from 'vite'
 import type { DepsOptimizationOptions, InlineConfig } from '../../types'
 
 export function resolveOptimizerConfig(testOptionc: DepsOptimizationOptions | undefined, viteOptions: DepOptimizationOptions | undefined, testConfig: InlineConfig) {
@@ -31,4 +31,40 @@ export function resolveOptimizerConfig(testOptionc: DepsOptimizationOptions | un
     }
   }
   return newConfig
+}
+
+export function deleteDefineConfig(viteConfig: ViteConfig) {
+  const defines: Record<string, any> = {}
+  if (viteConfig.define) {
+    delete viteConfig.define['import.meta.vitest']
+    delete viteConfig.define['process.env']
+    delete viteConfig.define.process
+  }
+  for (const key in viteConfig.define) {
+    const val = viteConfig.define[key]
+    let replacement: any
+    try {
+      replacement = typeof val === 'string' ? JSON.parse(val) : val
+    }
+    catch {
+      // probably means it contains reference to some variable,
+      // like this: "__VAR__": "process.env.VAR"
+      continue
+    }
+    if (key.startsWith('import.meta.env.')) {
+      const envKey = key.slice('import.meta.env.'.length)
+      process.env[envKey] = replacement
+      delete viteConfig.define[key]
+    }
+    else if (key.startsWith('process.env.')) {
+      const envKey = key.slice('process.env.'.length)
+      process.env[envKey] = replacement
+      delete viteConfig.define[key]
+    }
+    else if (!key.includes('.')) {
+      defines[key] = replacement
+      delete viteConfig.define[key]
+    }
+  }
+  return defines
 }
